feat: relacionando as respostas ao seu autor

+++ src/main/java/com/lukaskardeck/forum_hub/service/AnswerService.java
@@ -1,7 +1,7 @@
 package com.lukaskardeck.forum_hub.service;
 
 import com.lukaskardeck.forum_hub.domain.Answer;
-import com.lukaskardeck.forum_hub.domain.Topic;
+import com.lukaskardeck.forum_hub.domain.User;
 import com.lukaskardeck.forum_hub.dto.answer.AnswerDetailsResponse;
 import com.lukaskardeck.forum_hub.dto.answer.CreateAnswerRequest;
 import com.lukaskardeck.forum_hub.dto.answer.UpdateAnswerRequest;
@@ -11,6 +11,8 @@ import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -22,13 +24,19 @@ public class AnswerService {
     @Autowired
     private TopicRepository topicRepository;
 
+    @Autowired
+    private AuthService authService;
+
     public AnswerDetailsResponse createAnswer(CreateAnswerRequest answerRequest) {
         var topicId = answerRequest.topicId();
         var topic = topicRepository.findById(topicId).orElseThrow(
                 () -> new EntityNotFoundException("Tópico com ID " + topicId + " não encontrado")
         );
 
-        var newAnswer = new Answer(answerRequest, topic);
+        var currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        var userAuthor = (User) authService.loadUserByUsername(currentUsername);
+
+        var newAnswer = new Answer(answerRequest, topic, userAuthor);
         answerRepository.save(newAnswer);
 
         return new AnswerDetailsResponse(newAnswer);
@@ -56,6 +64,7 @@ public class AnswerService {
                 () -> new EntityNotFoundException("Resposta com ID " + idAnswer + " não encontrada")
         );
 
+        validateAuthor(answer);
         answer.update(updateRequest);
         answerRepository.flush();
 
@@ -68,7 +77,18 @@ public class AnswerService {
                 () -> new EntityNotFoundException("Resposta com ID " + id + " não encontrada")
         );
 
+        validateAuthor(answer);
         answerRepository.delete(answer);
+    }
+
+
+    private void validateAuthor(Answer answer) {
+        var currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        var answerAuthorUsername = answer.getAuthor().getUsername();
+
+        if (!answerAuthorUsername.equals(currentUsername)) {
+            throw new AccessDeniedException("Você não tem permissão para modificar esta resposta.");
+        }
     }
 }
